requirements.txt
psutil
python-dotenv
python-telegram-bot
requests